fixed dtype bugs and mdp.test() returns all passed!

+++ mdp/test/test_BasisFunctionNode.py
@@ -13,7 +13,7 @@ def test_basisfn_node():
     bfn = BasisFunctionNode(basis_name='indicator', lims=lims)
 
     for i in xrange(lims[1][0]):
-        inp = mdp.numx.array([[i]])
+        inp = mdp.numx.array([[i]]).astype('float')
         out = bfn(inp)
         exp_out = mdp.numx.zeros(lims[1][0] - lims[0][0] + 1)[None, :]
         exp_out[0, i] = 1.
